Part 30 From Insights to Visibility Monitoring your Market Intelligence Agent with ADOT/AgentCode/requirements.txt
boto3>=1.34.131
flask>=3.0.0
bedrock-agentcore
aws-opentelemetry-distro[otlp]
